feat: add claims disclosure screen
