+++ ts/features/itwallet/__mocks__/presentationClaims.json
@@ -1,0 +1,46 @@
+{
+  "requiredClaims": [
+    {
+      "claim": {
+        "id": "fiscal_code",
+        "label": "Codice fiscale",
+        "value": "RSSMRA70A01H501S"
+      },
+      "source": "IPZS"
+    },
+    {
+      "claim": {
+        "id": "name",
+        "label": "Nome",
+        "value": "Mario"
+      },
+      "source": "IPZS"
+    },
+    {
+      "claim": {
+        "id": "surname",
+        "label": "Cognome",
+        "value": "Rossi"
+      },
+      "source": "IPZS"
+    }
+  ],
+  "optionalClaims": [
+    {
+      "claim": {
+        "id": "birthdate",
+        "label": "Data di nascita",
+        "value": "01/01/1970"
+      },
+      "source": "IPZS"
+    },
+    {
+      "claim": {
+        "id": "city",
+        "label": "Città",
+        "value": "Roma"
+      },
+      "source": "IPZS"
+    }
+  ]
+}
