+++ ts/features/itwallet/common/utils/itwMocksUtils.ts
@@ -3,6 +3,8 @@ import eid from "../../__mocks__/eid.json";
 import mdl from "../../__mocks__/mdl.json";
 import statusAttestation from "../../__mocks__/statusAttestation.json";
 import ts from "../../__mocks__/ts.json";
+import presentationClaims from "../../__mocks__/presentationClaims.json";
+import { type DisclosureClaim } from "../components/ItwClaimsDisclosure";
 import { ParsedStatusAttestation, StoredCredential } from "./itwTypesUtils";
 
 export const ISSUER_MOCK_NAME = "Istituto Poligrafico e Zecca dello Stato";
@@ -26,4 +28,9 @@ export const ItwStoredCredentialsMocks = {
 
 export const ItwStatusAttestationMocks = {
   mdl: statusAttestation as ParsedStatusAttestation
+};
+
+export const ItwRemotePresentationClaimsMock = {
+  required: presentationClaims.requiredClaims as Array<DisclosureClaim>,
+  optional: presentationClaims.optionalClaims as Array<DisclosureClaim>
 };
